Precision Oncology-Based Cancer Therapeutics
Precision Oncology-Based Targeting of Signaling Pathways 
toward a Holistic Approach to Cancer Therapeutics

The genetic changes appearing in the information system of the cell that program its unregulated growth and proliferation gradually lead to cancer manifestation, and the treatment options need to be guided accordingly. The critical roles played by some of the molecules associated with the signaling pathways and cell microenvironment that often induce tumorigenesis and metastasis have been described precisely in recent years based on findings of the human genome project and other related initiatives undertaken afterward to thoroughly understand the molecular basis of cancer cell behaviors. It is to rely upon the genomic study of cancer cells to fully understand the prognosis and pathways involved in disease progression to selectively target them for a cure. Furthermore, patients with the same cancer types often respond differently to cancer therapies, indicating the need for a patient-specific treatment regimen for cancer. In this direction, precision oncology, defined as the molecular profiling of tumors to identify targetable alterations for custom-tailored personalized treatment, is gaining ground as a potential means of cancer treatment and has started influencing the ways cancer has been treated in clinics. This article intends to comprehensively elucidate the foundations and frontiers of precision oncology in the context of recent advances in cancer genomics and single-cell technologies for assessing its scope and importance in the realization of a proper cure for cancer.